[YOURLS] Move the Python API into api.py
